requirements.txt
fastapi
uvicorn
pdfminer.six
python-docx
spacy
httpx
supabase
pydantic
